requirements.txt
discord~=2.3.2
SpeechRecognition==3.9.0
google_cloud_speech~=2.26.0
google-generativeai~=0.5.2
frozenlist==1.3.3
py-cord==2.3.2
pycparser==2.21
python-dotenv==0.21.1